+++ src/onegov.people/onegov/people/__init__.py
@@ -1,4 +1,4 @@
-from onegov.people.collection import PersonCollection
+from onegov.people.collections import PersonCollection
 from onegov.people.models import Person
 
 __all__ = [