+++ src/onegov.people/onegov/people/collections/__init__.py
@@ -1,0 +1,6 @@
+from onegov.people.collections.people import PersonCollection
+
+
+__all__ = (
+    'PersonCollection',
+)