+++ src/onegov.people/onegov/people/models/__init__.py
@@ -1,0 +1,6 @@
+from onegov.people.models.person import Person
+
+
+__all__ = (
+    'Person',
+)